user_api/requirements.txt
Flask>=1.1.1
Flask-Cors>=3.0.8
python-dotenv>=0.10.3
Flask-SQLAlchemy>=2.4.0
SQLAlchemy>=1.3.8
pymysql>=0.9.3
flask-marshmallow>=0.10.1
marshmallow-sqlalchemy>=0.19.0
pymemcache>=2.2.2
pytest>=5.1.2
flask-swagger-ui>=3.20.9
cryptography>=2.6.1